requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.5
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.1
environs==11.0.0
idna==3.7
marshmallow==3.21.3
numpy==2.0.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
